delete trait
